* upstream/main: (1565 commits)
  Enable journal information fetcher directly in popup (#10598)
  Make generate button wider (#10588)
  make openRewrite stable again
  Rename cleanup_pr.yml to cleanup-pr.yml
  Changelog
  Fix failing fetcher tests (#10613)
  ShortDoi
  Fix CHANGELOG.md
  Bump commons-cli:commons-cli from 1.5.0 to 1.6.0 (#10607)
  Fix issue #9306: Move "Open only one instance of JabRef" preference option to somewhere else (#10602)
  Update README.md (#10604)
  Bump me.champeau.jmh from 0.7.1 to 0.7.2
  Bump org.beryx.jlink from 3.0.0 to 3.0.1
  Bump com.dlsc.gemsfx:gemsfx from 1.82.0 to 1.84.0
  Bump org.apache.logging.log4j:log4j-to-slf4j from 2.21.0 to 2.21.1
  Synchronize scrollbars in the change resolver dialog (#10587)
  Add button for a user to reset the cite command to the default value. (#10580)
  openrerwrite
  Update .github/PULL_REQUEST_TEMPLATE.md
  Update PULL_REQUEST_TEMPLATE.md
  ...

# Conflicts:
#	.devcontainer/devcontainer.json
#	build.gradle
#	src/main/java/org/jabref/gui/DefaultInjector.java
#	src/main/java/org/jabref/gui/libraryproperties/general/GeneralProperties.fxml
#	src/main/java/org/jabref/gui/libraryproperties/general/GeneralPropertiesView.java
#	src/main/java/org/jabref/gui/libraryproperties/general/GeneralPropertiesViewModel.java
#	src/main/java/org/jabref/logic/exporter/MetaDataSerializer.java
#	src/main/java/org/jabref/logic/importer/util/MetaDataParser.java
#	src/main/java/org/jabref/model/metadata/MetaData.java
#	src/main/java/org/jabref/preferences/JabRefPreferences.java

+++ src/main/java/org/jabref/logic/journals/JournalAbbreviationLoader.java
@@ -1,21 +1,31 @@
 package org.jabref.logic.journals;
 
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
+import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/**
+ * <p>
+ *   This class loads abbreviations from a CSV file and stores them into a MV file ({@link #readAbbreviationsFromCsvFile(Path)}
+ *   It can also create an {@link JournalAbbreviationRepository} based on an MV file ({@link #loadRepository(JournalAbbreviationPreferences)}.
+ * </p>
+ * <p>
+ *   Abbreviations are available at <a href="https://github.com/JabRef/abbrv.jabref.org/">https://github.com/JabRef/abbrv.jabref.org/</a>.
+ * </p>
+ */
 public class JournalAbbreviationLoader {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(JournalAbbreviationLoader.class);
 
-    public static List<Abbreviation> readJournalListFromFile(Path file) throws IOException {
-        LOGGER.debug(String.format("Reading journal list from file %s", file));
+    public static Collection<Abbreviation> readAbbreviationsFromCsvFile(Path file) throws IOException {
+        LOGGER.debug("Reading journal list from file {}", file);
         AbbreviationParser parser = new AbbreviationParser();
         parser.readJournalListFromFile(file);
         return parser.getAbbreviations();
@@ -23,14 +33,20 @@ public class JournalAbbreviationLoader {
 
     public static JournalAbbreviationRepository loadRepository(JournalAbbreviationPreferences journalAbbreviationPreferences) {
         JournalAbbreviationRepository repository;
+
         // Initialize with built-in list
-        try {
-            Path tempDir = Files.createTempDirectory("jabref-journal");
-            Path tempJournalList = tempDir.resolve("journalList.mv");
-            Files.copy(JournalAbbreviationRepository.class.getResourceAsStream("/journals/journalList.mv"), tempJournalList);
-            repository = new JournalAbbreviationRepository(tempJournalList);
-            tempDir.toFile().deleteOnExit();
-            tempJournalList.toFile().deleteOnExit();
+        try (InputStream resourceAsStream = JournalAbbreviationRepository.class.getResourceAsStream("/journals/journal-list.mv")) {
+            if (resourceAsStream == null) {
+                LOGGER.warn("There is no journal-list.mv. We use a default journal list");
+                repository = new JournalAbbreviationRepository();
+            } else {
+                Path tempDir = Files.createTempDirectory("jabref-journal");
+                Path tempJournalList = tempDir.resolve("journal-list.mv");
+                Files.copy(resourceAsStream, tempJournalList);
+                repository = new JournalAbbreviationRepository(tempJournalList);
+                tempDir.toFile().deleteOnExit();
+                tempJournalList.toFile().deleteOnExit();
+            }
         } catch (IOException e) {
             LOGGER.error("Error while copying journal list", e);
             return null;
@@ -38,13 +54,15 @@ public class JournalAbbreviationLoader {
 
         // Read external lists
         List<String> lists = journalAbbreviationPreferences.getExternalJournalLists();
-        if (!(lists.isEmpty())) {
+        // might produce NPE in tests
+        if (lists != null && !(lists.isEmpty())) {
+            // reversing ensures that the latest lists overwrites the former one
             Collections.reverse(lists);
             for (String filename : lists) {
                 try {
-                    repository.addCustomAbbreviations(readJournalListFromFile(Path.of(filename)));
+                    repository.addCustomAbbreviations(readAbbreviationsFromCsvFile(Path.of(filename)));
                 } catch (IOException e) {
-                    LOGGER.error(String.format("Cannot read external journal list file %s", filename), e);
+                    LOGGER.error("Cannot read external journal list file {}", filename, e);
                 }
             }
         }
@@ -52,6 +70,6 @@ public class JournalAbbreviationLoader {
     }
 
     public static JournalAbbreviationRepository loadBuiltInRepository() {
-        return loadRepository(new JournalAbbreviationPreferences(Collections.emptyList(), StandardCharsets.UTF_8, true));
+        return loadRepository(new JournalAbbreviationPreferences(Collections.emptyList(), true));
     }
 }
